comment out bfloat16

+++ app.py
@@ -9,7 +9,9 @@ def init():
 
     device = 0 if torch.cuda.is_available() else -1
     tokenizer = T5Tokenizer.from_pretrained("google/flan-t5-xl")
-    model = T5ForConditionalGeneration.from_pretrained("google/flan-t5-xl", device_map="auto", torch_dtype=torch.bfloat16)
+    model = T5ForConditionalGeneration.from_pretrained("google/flan-t5-xl", device_map="auto"
+                                                       # torch_dtype=torch.bfloat16
+    )
 
     input_text = "translate English to German: How old are you?"
     input_ids = tokenizer(input_text, return_tensors="pt").input_ids.to("cuda")
